Technology clubs clustering simulation
There are some papers in the economics literature on technology clubs which assume that if the clusters found really reflect a true structure in the data, a classifier will be able to make good predictions (high accuracy) about the cluster’s label.  This computer simulation shows that this assumption is incorrect.